renamed tests to use UnitTest suffix
