+++ spring-core/src/test/java/com/baeldung/di/AuthenticationServiceAnnotationConfigUnitTest.java
@@ -10,7 +10,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 import static org.junit.Assert.*;
 
 @RunWith(SpringRunner.class)
-public class AuthenticationServiceAnnotationConfigTest {
+public class AuthenticationServiceAnnotationConfigUnitTest {
     @Autowired
     private AuthenticationService authenticationService;
