+++ spring-core/src/test/java/com/baeldung/di/AuthenticationServiceXmlConfigUnitTest.java
@@ -6,17 +6,17 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.*;
 
 @RunWith(SpringRunner.class)
 @ContextConfiguration(locations = {"classpath:application-context.xml"})
-public class UserServiceXmlConfigTest {
+public class AuthenticationServiceXmlConfigUnitTest {
     @Autowired
-    private UserService userService;
+    private AuthenticationService authenticationService;
 
     @Test
-    public void givenApplicationContext_whenUserServiceIsInstantiated_thenDependenciesShouldNotBeNull() {
-        assertNotNull(userService);
-        assertNotNull(userService.getUserRepository());
+    public void givenApplicationContext_whenAuthenticationServiceIsInstantiated_thenDependenciesShouldNotBeNull() {
+        assertNotNull(authenticationService);
+        assertNotNull(authenticationService.getUserService());
     }
 }
